backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sarif-om>=1.0.4",
    "sqlalchemy>=2.0.41",
    "strands-agents>=1.1.0",
    "strands-agents-tools>=0.2.2",
    "temporalio>=1.14.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[build-system]
requires=["hatchling"]
build-backend="hatchling.build"

[tool.hatch.build.targets.wheel]
packages=["temporal_"]

[tool.pytest.ini_options]
testpaths=["temporal_"]
